+++ src/app/mypage/visitorList/ApplicationFormDialog.tsx
@@ -5,7 +5,7 @@ import {
   CompanyMst,
   IndividualSalesResult,
   ProductMst,
-} from "../types";
+} from "../../types";
 import Button from "@mui/material/Button";
 import Dialog from "@mui/material/Dialog";
 import DialogActions from "@mui/material/DialogActions";
