+++ src/app/mypage/visitorList/VisitorList.tsx
@@ -11,7 +11,7 @@ import {
   CompanyMst,
   IndividualSalesResult,
   ProductMst,
-} from "../types";
+} from "../../types";
 import EditIcon from "@mui/icons-material/Edit";
 import DeleteIcon from "@mui/icons-material/Delete";
 import IconButton from "@mui/material/IconButton";
